swipe images

+++ vite.config.ts
@@ -3,7 +3,7 @@ import { fileURLToPath, URL } from 'node:url'
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
 import git from 'git-rev-sync'
-process.env.VITE_GIT_COMMIT_HASH = git.short('./.git')
+process.env.VITE_GIT_COMMIT_HASH = git.long('./.git')
 
 // https://vitejs.dev/config/
 export default defineConfig({
